Change Put(box) to place box on grid
Before, Put() expected the box to already have the correct x,y origin to be
placed on the grid. Now, a canonized box with x,y=0 can be taken as input
and the origin is calculated within Put().
This makes sense as Put() will be fed from store where only canonized boxes
lie.