Rename PutOnGrid to SetOrigin and remove original
The original SetOrigin(a,b) really just set box.x=a and box.y=b regardless of
input. The PutOnGrid function basically did the same, but performed various
checks before. If all checks passed it just called SetOrigin().

This indirection and naming as 'PutOnGrid' leads to confusion of the actual
function and thus the old SetOrigin() is deleted and PutOnGrid() renamed as
SetOrigin().

All Tests passed.