Move sorting of freegrid from Put() to Update()
Old version: the three elements (bottom, right, bottomRight) that are
generated when a box is placed and the free grid is split where directly
sorted in Put().
Later they were appended to the FreeGrid. But this resulted in an unsorted
FreeGrid.

New Version: the total FreeGrid array after new elements are appended will
be sorted. This increased the speed and the profit.